requirements.txt
rich
python-dotenv
openai
